wire: move GetUtreexoSummaryHeights
GetUtreexoSummaryHeights is renamed to GetUtreexoTTLHeights and moved
to msggetutreexottls.go as we're no longer using msg getutreexosummaries
but still need the logic for getting heights.
