+++ wire/msggetutreexottls_test.go
@@ -6,6 +6,7 @@ package wire
 
 import (
 	"bytes"
+	"reflect"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -97,5 +98,67 @@ func TestMsgGetUtreexoTTLsEncode(t *testing.T) {
 		require.Equal(t, testCase.version, afterMsg.Version)
 		require.Equal(t, testCase.startHeight, afterMsg.StartHeight)
 		require.Equal(t, testCase.maxExponent, afterMsg.MaxReceiveExponent)
+	}
+}
+
+func TestGetUtreexoTTLHeights(t *testing.T) {
+	testCases := []struct {
+		height          int32
+		bestHeight      int32
+		exponent        uint8
+		expectedHeights []int32
+	}{
+		{
+			height:          0,
+			bestHeight:      2,
+			exponent:        1,
+			expectedHeights: []int32{0, 1},
+		},
+
+		{
+			height:          1,
+			bestHeight:      1,
+			exponent:        1,
+			expectedHeights: []int32{1},
+		},
+
+		{
+			height:          1,
+			bestHeight:      1,
+			exponent:        0,
+			expectedHeights: []int32{1},
+		},
+
+		{
+			height:          8,
+			bestHeight:      20,
+			exponent:        3,
+			expectedHeights: []int32{8, 9, 10, 11, 12, 13, 14, 15},
+		},
+
+		{
+			height:          8,
+			bestHeight:      20,
+			exponent:        3,
+			expectedHeights: []int32{8, 9, 10, 11, 12, 13, 14, 15},
+		},
+
+		{
+			height:          9,
+			bestHeight:      20,
+			exponent:        3,
+			expectedHeights: []int32{9, 10, 11, 12, 13, 14, 15},
+		},
+	}
+
+	for _, testCase := range testCases {
+		got, err := GetUtreexoTTLHeights(testCase.height, testCase.bestHeight, testCase.exponent)
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		if !reflect.DeepEqual(testCase.expectedHeights, got) {
+			t.Fatalf("expected %v, got %v", testCase.expectedHeights, got)
+		}
 	}
 }
